tests/wpt/mozilla/tests/mozilla/bluetooth/requestDevice/two-filters.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';
test(function() {
    window.testRunner.setBluetoothMockDataSet('GlucoseHeartRateAdapter');
    var glucose_device = window.navigator.bluetooth.requestDevice(
        { filters: [{services: ['battery_service']},
                    {services: ['heart_rate']}]}
    );
    assert_equals(glucose_device.name, 'Heart Rate Device');
}, 'An extra filter doesn\'t prevent matching.');
</script>
